Added openal soft libraries. Fixed a bug in displaying multiple lines of text through one object. Added cubemap shadow mapping. Added bloom to materials.

+++ Core/Font.h
@@ -14,7 +14,7 @@ namespace Core
 		float SpacingWidth;
 		float SpacingHeight;
 
-		Font(Core::Texture* texture, float charWidth, float charHeight, int charsPerRow, float spacingWidth, int spacingHeight);
+		Font(Core::Texture* texture, float charWidth, float charHeight, int charsPerRow, float spacingWidth, float spacingHeight);
 		~Font();
 	};
 }
